refactor(lotto): separate classes and methods
Added enum classes:
-money error messages
-number error messages
